requirements.txt
Django==5.0.14
djangorestframework==3.15.2
psycopg[binary]==3.3.3
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2
razorpay==1.4.2
pytest==8.3.5
pytest-django==4.11.1
